refactor: move useCanEdit to src/hooks
